+++ src/main/java/com/trabajointegrador/PauloPintos_MelisaFerraris/service/IOdontologoService.java
@@ -1,15 +1,19 @@
 package com.trabajointegrador.PauloPintos_MelisaFerraris.service;
 
 
-import com.trabajointegrador.PauloPintos_MelisaFerraris.Dto.OdontologoDto;
+import com.trabajointegrador.PauloPintos_MelisaFerraris.dto.OdontologoDto;
 import com.trabajointegrador.PauloPintos_MelisaFerraris.entity.Odontologo;
 
 import java.util.List;
 
 public interface IOdontologoService {
-    OdontologoDto buscarOdontologoPorId(int id);
+    OdontologoDto buscarOdontologoPorId(Long id);
+
     List<OdontologoDto> listarOdontologos();
+
     OdontologoDto registrarOdontologo(Odontologo odontologo);
+
     OdontologoDto actualizarOdontologo(Odontologo odontologo);
-    void eliminarOdontologo(int id);
+
+    void eliminarOdontologo(Long id);
 }
